+++ raiden-ts/src/raiden.ts
@@ -806,10 +806,16 @@ export class Raiden {
     const tokenNetwork = this.state.tokens[token];
     assert(tokenNetwork, ErrorCodes.RDN_UNKNOWN_TOKEN_NETWORK, this.log.info);
 
-    const decodedValue = decode(UInt(32), value);
-    const paymentId = options.paymentId ? decode(UInt(8), options.paymentId) : makePaymentId();
-    const paths = options.paths ? decode(Paths, options.paths) : undefined;
-    const pfs = options.pfs ? decode(PFS, options.pfs) : undefined;
+    const decodedValue = decode(UInt(32), value, ErrorCodes.DTA_INVALID_AMOUNT, this.log.info);
+    const paymentId = options.paymentId
+      ? decode(UInt(8), options.paymentId, ErrorCodes.DTA_INVALID_PAYMENT_ID, this.log.info)
+      : makePaymentId();
+    const paths = !options.paths
+      ? undefined
+      : decode(Paths, options.paths, ErrorCodes.DTA_INVALID_PATH, this.log.info);
+    const pfs = !options.pfs
+      ? undefined
+      : decode(PFS, options.pfs, ErrorCodes.DTA_INVALID_PFS, this.log.info);
 
     assert(
       options.secret === undefined || Secret.is(options.secret),